requirements.txt
Django>=3.0,<4.0
psycopg2-binary
djangorestframework
django-filter